rename: rename Coins to RandomGenerator and move utils package
